+++ packages/react-server/src/server/server.ts
@@ -1,11 +1,8 @@
 import 'isomorphic-fetch';
 import {Server} from 'http';
-import {join} from 'path';
-import {existsSync} from 'fs';
 import Koa, {Context} from 'koa';
 import compose from 'koa-compose';
 import mount from 'koa-mount';
-import {middleware as sewingKitMiddleware} from '@shopify/sewing-kit-koa';
 import {createRender, RenderFunction} from '../render';
 import {requestLogger} from '../logger';
 import {metricsMiddleware as metrics} from '../metrics';
@@ -29,33 +26,19 @@ interface Options {
 export function createServer(options: Options): Server {
   const {port, assetPrefix, render, serverMiddleware, ip} = options;
   const app = new Koa();
-  const manifestPath = getManifestPath(process.cwd());
 
   app.use(mount('/services/ping', ping));
 
   app.use(requestLogger);
   app.use(metrics);
-  app.use(sewingKitMiddleware({assetPrefix, manifestPath}));
 
   if (serverMiddleware) {
     app.use(compose(serverMiddleware));
   }
 
-  app.use(createRender(render));
+  app.use(createRender(render, {assetPrefix}));
 
   return app.listen(port || 3000, () => {
     logger.log(`started react-server on ${ip}:${port}`);
   });
-}
-
-function getManifestPath(root: string) {
-  const gemFileExists = existsSync(join(root, 'Gemfile'));
-  if (!gemFileExists) {
-    return;
-  }
-
-  // eslint-disable-next-line no-process-env
-  return process.env.NODE_ENV === 'development'
-    ? `tmp/sewing-kit/sewing-kit-manifest.json`
-    : `public/bundles/sewing-kit-manifest.json`;
 }
